fix: add DatasetSplitName in enums.py
