+++ tests/models/dataset/fixtures/dataset_collection_fixtures.py
@@ -3,7 +3,7 @@ from typing import Callable
 import pytest
 from picsellia.types.enums import InferenceType
 
-from src.models.dataset.dataset_split_name import DatasetSplitName
+from src.enums import DatasetSplitName
 from src.models.dataset.dataset_collection import DatasetCollection
 from tests.steps.fixtures.dataset_version_fixtures import DatasetTestMetadata
