+++ tests/steps/data_preparation/utils/test_dataset_organizer.py
@@ -3,7 +3,7 @@ from typing import Callable
 
 from picsellia.types.enums import InferenceType
 
-from src.models.dataset.dataset_split_name import DatasetSplitName
+from src.enums import DatasetSplitName
 from tests.steps.fixtures.dataset_version_fixtures import DatasetTestMetadata
